Wagon class bug fix and Initial game set.
The constructor in the wagon class was fixed so that the leader is entered into the persons array and the counter for the followers are shifted.

The main driver file now starts off the game by welcoming the user to the game and calling the wagon constructor.
After all of the members are entered, the main file also informs the user about the store and takes them into the store simulator.