Constructor fix for Wagon class
The wagon class will now add all of the supplies to the materials array where as before when they were created but never added. Each element in which they are added to is hardcoded in becuase they are constant and will non will be added or taken away throughout the game.

the money data member has also been changed to a float type.