New Auxiliary file and bug fixes
A new file was added to house all of the helper functions used throughout the program. Currently the file only has the amount function, which asks the user to enter a value and then checks the input for char and casts into an int should it only contain digits, and the clear function which asks the user to press enter to clear the screen and move to the next section.

The wagon class file will include the auxiliary file for the rest of the program to use and an extra member function was added. The wagon class will now contain a new getter for materials which has an int parameter rather than an string parameter. This is to fix the medkit/food bug from the last commit.

The Oregon_Trail_Driver file now used the new member function from the wagon class for the medkit display cost. The store function will also utilize the new amount function when asking the user for amount in supplies in order to sort through edgecases easier.