Store bug fix and new functions
The store can now differentiate between food and medKits for sure this time. This was fixed by removing the getter function with a string as a parameter and strictly using the one with the index for the parameter when calling materials anywhere.

Due to setters from the supply and WagonGang class not working with the getters from the Wagon class, new setter were needed in the wagon class so that objects called through the wagon can be changed with a single member function.

A save and load function were also added to the main driver file. Though they are not yet functional, their bare bones are layed out for futher development