Menu Skeleton added and Date class modification
A menu has been added to the display function for the user to choose the next course of action. Each choice will lead to a seperate function which has also been laid out in the diver file.

the time class now will only accept dates within 30 or 31 days depending on the month.