Menu function added with functionality in continue and stop. New milestone class. minor fixes
The driver file now has functions which allow the user to continue on the trail or stop and rest. As of now the quit and hunt do not yet work.In the continue function, the program checks whether or not the user is at a milestone now and gives the user choice based on what kind of marker they are at.

There is also a new milestone class which stores and modifies information on each land mark. there are getters for each data member and one setter which shifts all the data members in their respective class up one for all of the getter functions. This class is ran through the Wagon calss where getters and setters also exist for a milestone object.

A random function has been added to the auxiliary file so that the miles travelled varies from 70 to 140 each turn.

The main file will now also check the players status each turn to see if they meet any requirements to end the game. This includes running out of food, having the leader die, or running out of oxen.