Shop function now works everywhere, new constructor in wagon class, misfourtones added
The shop function has been modified to be used multiple times through out the trail with prices increasing depending on where the shop is accecced. The shop now utilized a seperate wagon object refered to as "cart". the cart will hold all the items the player wishes to purchase and then addes them all to the actual wagon when the shop is exited.

The new wagon constructor only initilizes supplies so that it can be used as a cart for the shop.

All misfortunes have been added to the class and each one has its respective consequences added